# Update .NET benchmark code - simplify and improve performance