--- a/2024/07/05/SimdHTML.sln
+++ b/2024/07/05/SimdHTML.sln
@@ -3,11 +3,9 @@ Microsoft Visual Studio Solution File, Format Version 12.00
 # Visual Studio Version 17
 VisualStudioVersion = 17.0.31903.59
 MinimumVisualStudioVersion = 10.0.40219.1
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "benchmark", "benchmark\benchmark.csproj", "{6C7744C5-AF3E-446A-8475-B03001989AB8}"
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "SimdHTML", "src\SimdHTML.csproj", "{3806E828-9AFF-47C3-854F-C8049DC2091E}"
 EndProject
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "SimdHTML", "src\SimdHTML.csproj", "{14D65B6D-24CD-4616-AAEF-ADBEB5E390A0}"
-EndProject
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "tests", "test\tests.csproj", "{BDAFE9A2-0CC2-4F72-919F-7D9A8D185EBE}"
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "benchmark", "benchmark\benchmark.csproj", "{A0993DE7-14A4-4206-B5D3-6FCD0B6D46C3}"
 EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
@@ -18,17 +16,13 @@ Global
 		HideSolutionNode = FALSE
 	EndGlobalSection
 	GlobalSection(ProjectConfigurationPlatforms) = postSolution
-		{6C7744C5-AF3E-446A-8475-A03001989AB8}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{6C7744C5-AF3E-446A-8475-A03001989AB8}.Debug|Any CPU.Build.0 = Debug|Any CPU
-		{6C7744C5-AF3E-446A-8475-A03001989AB8}.Release|Any CPU.ActiveCfg = Release|Any CPU
-		{6C7744C5-AF3E-446A-8475-A03001989AB8}.Release|Any CPU.Build.0 = Release|Any CPU
-		{14D65B6D-24CD-4616-AAEF-BDBEB5E390A0}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{14D65B6D-24CD-4616-AAEF-BDBEB5E390A0}.Debug|Any CPU.Build.0 = Debug|Any CPU
-		{14D65B6D-24CD-4616-AAEF-BDBEB5E390A0}.Release|Any CPU.ActiveCfg = Release|Any CPU
-		{14D65B6D-24CD-4616-AAEF-BDBEB5E390A0}.Release|Any CPU.Build.0 = Release|Any CPU
-		{BDAFE9A2-0CC2-4F72-919F-5D9A8D185EBE}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{BDAFE9A2-0CC2-4F72-919F-5D9A8D185EBE}.Debug|Any CPU.Build.0 = Debug|Any CPU
-		{BDAFE9A2-0CC2-4F72-919F-5D9A8D185EBE}.Release|Any CPU.ActiveCfg = Release|Any CPU
-		{BDAFE9A2-0CC2-4F72-919F-5D9A8D185EBE}.Release|Any CPU.Build.0 = Release|Any CPU
+		{3806E828-9AFF-47C3-854F-C8049DC2091E}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{3806E828-9AFF-47C3-854F-C8049DC2091E}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{3806E828-9AFF-47C3-854F-C8049DC2091E}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{3806E828-9AFF-47C3-854F-C8049DC2091E}.Release|Any CPU.Build.0 = Release|Any CPU
+		{A0993DE7-14A4-4206-B5D3-6FCD0B6D46C3}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{A0993DE7-14A4-4206-B5D3-6FCD0B6D46C3}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{A0993DE7-14A4-4206-B5D3-6FCD0B6D46C3}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{A0993DE7-14A4-4206-B5D3-6FCD0B6D46C3}.Release|Any CPU.Build.0 = Release|Any CPU
 	EndGlobalSection
 EndGlobal