This pull-request includes the following changes. For better reviewing experience, consider using "ignore whitespace" option. Thanks!

#### Refactor `Neon64Match` into an idiomatic enumerator 

C# supports regular (implementing `IEnumerable<T>` interface) and duck-typed enumerators.
Therefore, instead of C++-style custom loop, a regular `foreach (var match in ...` can be used instead.
Additionally, make scanning methods and state mutation ones separate - instance methods on structs take them
by reference, and this tends to, sometimes, make the compiler output worse because it forces struct contents
onto the stack instead of keeping them in registers at all times. This is an area where .NET has significantly closed the gap
with LLVM and GCC starting with version 8, but it is more fragile still, so we have to help it out occasionally.

Notably, `stackalloc` is a very special keyword which inserts stack cookie to guard against stack overrun with a
fail-fast path should that happen. It also prevents or severely limits method inlining. Because `CarefulUpdate` was
an instance method, it means that an un-inlineable method was taking `Neon64Match` by reference and was
forcing it to be spilled onto the stack which was degrading the quality of the loop significantly.

#### Make iterator-like types in `SimdHTML.cs` structs

This is important because, unlike OpenJDK, .NET has weaker escape analysis (it *just* gained such capability properly in .NET 9 preview). Of course .NET is a much lower level platform so historically this has almost never been an issue as structs vs classes distinction always clearly communicated how the type is supposed to be handled.

Just for the sake of clarity, .NET classes are strictly reference types and are more akin to `Box<TClassLayout>` as the Rust names it, they are distinct from C++ classes which might cause confusion.

This means that a state of an iterator that is a class will have to be reloaded per-iteration, and the current rendition of escape analysis is unable to see that it is entirely thread-private and remove reloading the data from/to heap.

#### Simplify benchmark code

I have opted to move out repeated `Speed` column code into a `Utils` project. Should you ever need it, you can simply `dotnet add reference path/to/folder/with/utils/project` and then annotate the benchmark fixture that you want to add it to. It would still look for `FileName` field as it does right now.

Also I made other cosmetic adjustments to reduce the line count - BenchmarkDotNet configurability is a bit of a mess, but the "happy-path" does not need as many LOC. I have also recreated .sln to make it correct and updated .gitignore.

Lastly, the project target was bumped to net9.0 (preview). If you are okay with it, please make sure to get the latest preview from https://dotnet.microsoft.com/en-us/download/dotnet/9.0. If not - please let me know.

#### Numbers

```
BenchmarkDotNet v0.13.12, macOS 15.0 (24A5289h) [Darwin 24.0.0]
Apple M1 Pro, 1 CPU, 8 logical and 8 physical cores
.NET SDK 9.0.100-preview.7.24367.22
  [Host]     : .NET 9.0.0 (9.0.24.36618), Arm64 RyuJIT AdvSIMD
  Job-WIRYJV : .NET 9.0.0 (9.0.24.36618), Arm64 RyuJIT AdvSIMD

IterationCount=6  LaunchCount=1  WarmupCount=3  
```

| Method            | FileName       | Mean        | Error    | StdDev   | Speed (GB/s) |
|------------------ |--------------- |------------:|---------:|---------:|------------- |
| SIMDHTMLScan      | data/data.html |  3,508.4 ns | 10.47 ns |  2.72 ns | 5.79         |
| NEONHTMLScan      | data/data.html |    875.0 ns |  1.68 ns |  0.60 ns | 23.22        |
| NEON64HTMLScan    | data/data.html |    734.9 ns |  0.97 ns |  0.25 ns | 27.65        |
| HTMLScan          | data/data.html | 15,666.6 ns | 98.12 ns | 25.48 ns | 1.30         |
| SearchValuesBench | data/data.html |  4,964.8 ns | 26.13 ns |  9.32 ns | 4.09         |

Sadly, this seems to ruin nice improvement over not unrolled variant in `NEONHTMLScan`. It appears currently implementation does not extract as much load throughput from M-series as theoretically achievable. Possible areas of exploration can be fast multi-vector table lookups that it supports, or making the implementation less fancy, I think. Current implementation seems to be generally inspired by how `movmask` works, and it's always a bit painful for patterns where SSE/AVX fit perfectly :)